schemadiff: split diff.go into per-concern files
Pure reorganization (no behavior change) to keep files focused as the engine
grows: difftypes.go (DiffType + Difference model), tables.go (diffTables and
table helpers), columns.go (diffColumns and column helpers), diff.go (Diff
orchestration, suppression, sort). Table tests moved to tables_test.go; shared
helpers and orchestration tests stay in diff_test.go. filter.go/columns_test.go
unchanged. 56 unit tests + live-PG integration test still green.
